unix timestamps do not work
the creation date is just the date, not the full time, hence why it gave me the 1970-01-01 error.

your database changes will not be used as they are incompatible with the framebit-styled database poktube now uses. @forkiesassds

the only use i would find for unix timestamps is if we were to do a gource video, which won't happen.

the structure is incompatible, it had to be fixed by myself.