DECLINING A VIDEO SHOULD NOT DELETE IT
For christ sakes. This does not delete the video. What if some spasctic faggot with the name of "Backyardigans 1999 Fuckboy UWU UTTP" basically just uploads 2gb worth of shit (yes, even though FFMPEG compresses it) and we're stuck with 300+ backyardigans shit that no one is able to watch (not just of cringe, but because it's stuck on the hard drive.)

ethier make the video accessible for admins, or just delete the video file but keep an entry for it.